- Added a "simulated" driver - Added a plc4c_system_add_driver to the "system" domain - Extended the hello_world.c to instantiate the "simulated" driver and register that at the "system".
