+++ sandbox/plc4c/drivers/CMakeLists.txt
@@ -19,3 +19,4 @@
 
 add_subdirectory(modbus)
 add_subdirectory(s7)
+add_subdirectory(simulated)
