+++ sandbox/plc4c/drivers/s7/CMakeLists.txt
@@ -17,12 +17,14 @@
   under the License.
 ]]
 
+include_directories("include" "../../api/include" "../../spi/include")
+
 # Add the sources
 file(GLOB sources "src/*.c")
 
 # Add the generated sources
 file(GLOB generatedSources "${PLC4C_ROOT_DIR}/target/generated-sources/plc4x/s7/src/main/c/*.c")
 
-add_library(plc4c-drivers-s7 ${sources} ${generatedSources})
+add_library(plc4c-driver-s7 ${sources} ${generatedSources})
 
-target_link_libraries (plc4c-drivers-s7 ${CMAKE_DL_LIBS})
+target_link_libraries (plc4c-driver-s7 ${CMAKE_DL_LIBS})
